added client creation to rpc exchange
