+++ crates/alator-exchange/src/rpc.rs
@@ -149,6 +149,30 @@ pub struct RPCExchange {
 }
 
 impl RPCExchange {
+    pub async fn build_exchange_client(client: tokio::io::DuplexStream, path: &str) -> Result<RPCExchange, Box<dyn std::error::Error>> {
+        let mut client = Some(client);
+        let channel = tonic::transport::Endpoint::try_from(path.to_owned())?
+            .connect_with_connector(tower::service_fn(move |_: tonic::transport::Uri| {
+                let client = client.take();
+
+                async move {
+                    if let Some(client) = client {
+                        Ok(client)
+                    } else {
+                        Err(std::io::Error::new(
+                            std::io::ErrorKind::Other,
+                            "Client already taken",
+                        ))
+                    }
+                }
+            }))
+            .await?;
+
+        let client = ExchangeClient::new(channel);
+        Ok(RPCExchange::new(client))
+    }
+    
+
     pub fn build_exchange_server(
         clock: Clock,
         source: DefaultPriceSource,
